requirements.txt
fastapi
uvicorn
motor
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1